Add files via upload
One common task for computers is to sort data. For example, people might want to see all their files on a computer sorted by size. Since sorting is a simple problem with many different possible solutions, it is often used to introduce the study of algorithms.Insertion Sort, a simple and intuitive sorting algorithm. We will first start with a nearly sorted list.